refactor: move services into services folder
