+++ src/main/java/com/cewar/services/GameService.java
@@ -1,8 +1,10 @@
-package com.cewar.game;
+package com.cewar.services;
 
 import java.util.HashMap;
 
 import org.springframework.stereotype.Service;
+
+import com.cewar.game.GameData;
 
 /**
  * A service to manage game objects
