+++ src/main/java/com/cewar/services/MyUserDetailsService.java
@@ -1,10 +1,11 @@
-package com.cewar.web.userdetails;
+package com.cewar.services;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Component;
 
-import com.cewar.repository.UserRepository;
+import com.cewar.repositories.UserRepository;
+import com.cewar.web.userdetails.User;
 
 /**
  * A User DAO (Data Access Object) helper class for DaoAuthenticationProvider.
